Added new models for keeping track of hardware lifecycle and support contracts.
A node now can has hardware lifecycle associated with it. It will be a one to one relationship.
A node now can be associated with one or many support contracts. 
A support contract can be associated with one or many nodes.

Updated permision filter in the application controller to return 401 http status for unauthorized actions



git-svn-id: svn://svn.code.sf.net/p/nventory/code/trunk@327 ee601400-a34b-0410-bf97-8c122c3a034e