We needed to add some exception catches in order to work on some of our test machines. No clue why some worked and others didn't, as it appears to happen regardless of the OS or hardware type.
I spoke with Jason briefly about these changes for the coreCount bits. Initially I had assumed that we were testing on older hardware and that, for whatever reason, they didn't register a core count (Jason pointed out that this happens in Linux on older CPUs in /proc/cpuinfo). Looking at what Windows tells me the CPU/Core count is, it looks like we're actually detecting the core as a CPU. I guess if I get some time later on we'll want to have it figure out the difference between a CPU and a Core and if it's a CPU grab the NumberOfCores value.

Anyway, this is really just a quick fix so we can get it working for our Windows hosts, we'll revisit this later and push more fixes as the need arises.



git-svn-id: svn://svn.code.sf.net/p/nventory/code/trunk@333 ee601400-a34b-0410-bf97-8c122c3a034e